Implementing partial runs:
I. Big refactor of the Spider.
Not using the CrawlSpider's features anymore. Loading the first page
and scraping the rest by ourselves now so that we have control
over the process and can stop whenever we want (partial run).

Added some date extraction to the parser in order to achieve that as
well.

New workflow:

                       >> requests to item's pages with parse_item_page as callback
Start page >> parse()
                       >> requests to itemlists' pages with parse_itemlist_page as callback

Then, the rest is unchanged: parse_itemlist_page >> requests with
parse_item_page as a callback.

Difference is, when the run is partial, during the ITEMLIST parsing, we
check the date of the items before even creating the requests to those
items' pages so that we save those requests because we already scraped
those guys.

II.  add_job now has more arguments (interval for partial vs. for full
runs)

III. Added the necessary new constants to config.py

IV. Big refactor of run_all_jobs

Additionally to previous refactor that was now forking jobs. We are now
taking into account whether we should run a full or a partial run
depending on the given intervals and on the last times we've run a full
run / partial run.

In the case of the partial run, passing to the spider the time it has to
scrap up to as an argument.

Also added some logging + error handling + logging in case of error
during the external processes. Still not quite perfect, though.